refactor: added templates and shorts
